+++ workspaces/core/src/index.ts
@@ -24,6 +24,8 @@ export { default as Loader } from './loader/Loader';
 export * from './loader/Loader';
 export { default as RadioButton } from './radioButton/RadioButton';
 export * from './radioButton/RadioButton';
+export { default as Select } from './select/Select';
+export * from './select/Select';
 export { default as TextInput } from './textInput/TextInput';
 export * from './textInput/TextInput';
 export { default as TextInputGroup } from './textInputGroup/TextInputGroup';