+++ workspaces/core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @heathmont/moon-core-tw
 
+## 9.7.0
+
+### Minor Changes
+
+- Add headless Table and Select
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-icons-tw@9.7.0
+
 ## 9.6.0
 
 ### Minor Changes